requirements.txt
pip
bilibili-api-python
matplotlib
pandas
aiohttp